Rename black.h to board_black.h
